+++ sdk/go/openstack/networking/subnet.go
@@ -158,7 +158,8 @@ func (r *Subnet) GatewayIp() *pulumi.StringOutput {
 	return (*pulumi.StringOutput)(r.s.State["gatewayIp"])
 }
 
-// An array of routes that should be used by devices
+// (**Deprecated** - use `openstack_networking_subnet_route_v2`
+// instead) An array of routes that should be used by devices
 // with IPs from this subnet (not including local subnet route). The host_route
 // object structure is documented below. Changing this updates the host routes
 // for the existing subnet.
@@ -260,7 +261,8 @@ type SubnetState struct {
 	// gateway of `.1` to be used. Changing this updates the gateway IP of the
 	// existing subnet.
 	GatewayIp interface{}
-	// An array of routes that should be used by devices
+	// (**Deprecated** - use `openstack_networking_subnet_route_v2`
+	// instead) An array of routes that should be used by devices
 	// with IPs from this subnet (not including local subnet route). The host_route
 	// object structure is documented below. Changing this updates the host routes
 	// for the existing subnet.
@@ -325,7 +327,8 @@ type SubnetArgs struct {
 	// gateway of `.1` to be used. Changing this updates the gateway IP of the
 	// existing subnet.
 	GatewayIp interface{}
-	// An array of routes that should be used by devices
+	// (**Deprecated** - use `openstack_networking_subnet_route_v2`
+	// instead) An array of routes that should be used by devices
 	// with IPs from this subnet (not including local subnet route). The host_route
 	// object structure is documented below. Changing this updates the host routes
 	// for the existing subnet.