+++ sdk/python/pulumi_openstack/compute/get_availability_zones.py
@@ -8,23 +8,20 @@ import pulumi
 import pulumi.runtime
 from .. import utilities, tables
 
-class GetGroupResult:
+class GetAvailabilityZonesResult:
     """
-    A collection of values returned by getGroup.
+    A collection of values returned by getAvailabilityZones.
     """
-    def __init__(__self__, domain_id=None, region=None, id=None):
-        if domain_id and not isinstance(domain_id, str):
-            raise TypeError('Expected argument domain_id to be a str')
-        __self__.domain_id = domain_id
+    def __init__(__self__, names=None, region=None, id=None):
+        if names and not isinstance(names, list):
+            raise TypeError('Expected argument names to be a list')
+        __self__.names = names
         """
-        See Argument Reference above.
+        The names of the availability zones, ordered alphanumerically, that match the queried `state`
         """
         if region and not isinstance(region, str):
             raise TypeError('Expected argument region to be a str')
         __self__.region = region
-        """
-        See Argument Reference above.
-        """
         if id and not isinstance(id, str):
             raise TypeError('Expected argument id to be a str')
         __self__.id = id
@@ -32,20 +29,17 @@ class GetGroupResult:
         id is the provider-assigned unique ID for this managed resource.
         """
 
-async def get_group(domain_id=None,name=None,region=None,opts=None):
+async def get_availability_zones(region=None,state=None,opts=None):
     """
-    Use this data source to get the ID of an OpenStack group.
-    
-    Note: This usually requires admin privileges.
+    Use this data source to get a list of availability zones from OpenStack
     """
     __args__ = dict()
 
-    __args__['domainId'] = domain_id
-    __args__['name'] = name
     __args__['region'] = region
-    __ret__ = await pulumi.runtime.invoke('openstack:identity/getGroup:getGroup', __args__, opts=opts)
+    __args__['state'] = state
+    __ret__ = await pulumi.runtime.invoke('openstack:compute/getAvailabilityZones:getAvailabilityZones', __args__, opts=opts)
 
-    return GetGroupResult(
-        domain_id=__ret__.get('domainId'),
+    return GetAvailabilityZonesResult(
+        names=__ret__.get('names'),
         region=__ret__.get('region'),
         id=__ret__.get('id'))