+++ sdk/python/pulumi_openstack/identity/get_endpoint.py
@@ -8,7 +8,7 @@ import pulumi
 import pulumi.runtime
 from .. import utilities, tables
 
-class GetEndpointResult(object):
+class GetEndpointResult:
     """
     A collection of values returned by getEndpoint.
     """
@@ -32,7 +32,7 @@ class GetEndpointResult(object):
         id is the provider-assigned unique ID for this managed resource.
         """
 
-async def get_endpoint(interface=None, region=None, service_id=None, service_name=None):
+async def get_endpoint(interface=None,region=None,service_id=None,service_name=None,opts=None):
     """
     Use this data source to get the ID of an OpenStack endpoint.
     
@@ -44,7 +44,7 @@ async def get_endpoint(interface=None, region=None, service_id=None, service_nam
     __args__['region'] = region
     __args__['serviceId'] = service_id
     __args__['serviceName'] = service_name
-    __ret__ = await pulumi.runtime.invoke('openstack:identity/getEndpoint:getEndpoint', __args__)
+    __ret__ = await pulumi.runtime.invoke('openstack:identity/getEndpoint:getEndpoint', __args__, opts=opts)
 
     return GetEndpointResult(
         region=__ret__.get('region'),