+++ sdk/python/pulumi_openstack/images/get_image.py
@@ -8,7 +8,7 @@ import pulumi
 import pulumi.runtime
 from .. import utilities, tables
 
-class GetImageResult(object):
+class GetImageResult:
     """
     A collection of values returned by getImage.
     """
@@ -101,7 +101,7 @@ class GetImageResult(object):
         id is the provider-assigned unique ID for this managed resource.
         """
 
-async def get_image(member_status=None, most_recent=None, name=None, owner=None, properties=None, region=None, size_max=None, size_min=None, sort_direction=None, sort_key=None, tag=None, visibility=None):
+async def get_image(member_status=None,most_recent=None,name=None,owner=None,properties=None,region=None,size_max=None,size_min=None,sort_direction=None,sort_key=None,tag=None,visibility=None,opts=None):
     """
     Use this data source to get the ID of an available OpenStack image.
     """
@@ -119,7 +119,7 @@ async def get_image(member_status=None, most_recent=None, name=None, owner=None,
     __args__['sortKey'] = sort_key
     __args__['tag'] = tag
     __args__['visibility'] = visibility
-    __ret__ = await pulumi.runtime.invoke('openstack:images/getImage:getImage', __args__)
+    __ret__ = await pulumi.runtime.invoke('openstack:images/getImage:getImage', __args__, opts=opts)
 
     return GetImageResult(
         checksum=__ret__.get('checksum'),