+++ sdk/python/pulumi_openstack/networking/get_network.py
@@ -8,7 +8,7 @@ import pulumi
 import pulumi.runtime
 from .. import utilities, tables
 
-class GetNetworkResult(object):
+class GetNetworkResult:
     """
     A collection of values returned by getNetwork.
     """
@@ -51,7 +51,7 @@ class GetNetworkResult(object):
         id is the provider-assigned unique ID for this managed resource.
         """
 
-async def get_network(description=None, external=None, matching_subnet_cidr=None, name=None, network_id=None, region=None, status=None, tags=None, tenant_id=None, transparent_vlan=None):
+async def get_network(description=None,external=None,matching_subnet_cidr=None,name=None,network_id=None,region=None,status=None,tags=None,tenant_id=None,transparent_vlan=None,opts=None):
     """
     Use this data source to get the ID of an available OpenStack network.
     """
@@ -67,7 +67,7 @@ async def get_network(description=None, external=None, matching_subnet_cidr=None
     __args__['tags'] = tags
     __args__['tenantId'] = tenant_id
     __args__['transparentVlan'] = transparent_vlan
-    __ret__ = await pulumi.runtime.invoke('openstack:networking/getNetwork:getNetwork', __args__)
+    __ret__ = await pulumi.runtime.invoke('openstack:networking/getNetwork:getNetwork', __args__, opts=opts)
 
     return GetNetworkResult(
         admin_state_up=__ret__.get('adminStateUp'),